Move clustering package to net.sansa.ml.spark dir
