+++ sansa-ml-spark/src/main/scala/net/sansa_stack/ml/spark/clustering/package.scala
@@ -1,5 +1,6 @@
-package net.sansa_stack.ml.spark.clustering.algorithms
+package net.sansa_stack.ml.spark
 
+import net.sansa_stack.ml.spark.clustering.algorithms._
 import net.sansa_stack.rdf.spark.io._
 import net.sansa_stack.rdf.spark.model._
 import org.apache.jena.graph.Triple
@@ -7,7 +8,7 @@ import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.SparkSession
 
 package object clustering {
-    object ClusteringAlgorithm extends Enumeration {
+  object ClusteringAlgorithm extends Enumeration {
     type ClusteringAlgorithm = Value
 
     val KMeans = Value("KMeans")
@@ -21,7 +22,7 @@ package object clustering {
   implicit class clusterT(input: RDD[Triple]) extends Enumeration {
     def cluster(name: ClusteringAlgorithm.Value): ClusterAlgo = name match {
       case ClusteringAlgorithm.KMeans => new Kmeans(input)
-      case ClusteringAlgorithm.DBSCAN => new DBSCAN(input)
+      case ClusteringAlgorithm.DBSCAN  => new DBSCAN(input)
       case ClusteringAlgorithm.PIC => new PIC(input)
       case ClusteringAlgorithm.BorderFlow => new BorderFlow(input.asGraph())
       case ClusteringAlgorithm.SilviaClustering => new SilviaClustering(input.asStringGraph())
@@ -29,4 +30,3 @@ package object clustering {
     }
   }
 }
-
