+++ sansa-ml-spark/src/test/scala/net/sansa_stack/ml/spark/clustering/algorithms/KmeansTest.scala
@@ -7,7 +7,7 @@ import net.sansa_stack.rdf.spark.model.graph._
 import org.apache.jena.riot.Lang
 import org.scalatest.FunSuite
 
-import net.sansa_stack.ml.spark.clustering.algorithms.clustering._
+import net.sansa_stack.ml.spark.clustering._
 
 class KmeansTest extends FunSuite with DataFrameSuiteBase {
   test("Kmeans.oneHotEncoding") {
